내비게이션 파라미터 전달 방식 변경

+++ app/src/main/java/com/drunkenboys/calendarun/ui/searchschedule/model/DateScheduleItem.kt
@@ -5,7 +5,7 @@ import java.text.SimpleDateFormat
 import java.util.*
 import java.util.Calendar.*
 
-data class DateScheduleItem(val schedule: Schedule, val onClick: (Int) -> Unit) {
+data class DateScheduleItem(val schedule: Schedule, val onClick: () -> Unit) {
 
     val duration: String = run {
         val startDateFormat = SimpleDateFormat("hh:mm", Locale.getDefault())
